Retrieve 404 when deleting an invalid company

+++ src/main/java/amazing/co/services/CompanyService.java
@@ -1,6 +1,7 @@
 package amazing.co.services;
 
 import amazing.co.exceptions.DuplicatedEntityException;
+import amazing.co.exceptions.EntityNotFoundException;
 import amazing.co.models.Company;
 import amazing.co.repositories.CompanyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,6 +28,13 @@ public class CompanyService {
     }
 
     public void delete(Long id) {
+        if (companyDoesNotExists(id)) {
+            throw new EntityNotFoundException("Company does not exist");
+        }
         companyRepository.deleteById(id);
+    }
+
+    private boolean companyDoesNotExists(Long id) {
+        return !companyRepository.existsById(id);
     }
 }
